#124 fix warning, return flecs::entity in entity_view::lookup and others

+++ flecs.c
@@ -398,7 +398,7 @@ void* _ecs_sparse_add(
     ((type*)_ecs_sparse_add(sparse, sizeof(type)))
 
 FLECS_API
-int32_t ecs_sparse_last_id(
+uint64_t ecs_sparse_last_id(
     const ecs_sparse_t *sparse);
 
 FLECS_API
@@ -10595,7 +10595,7 @@ void* _ecs_sparse_add(
     return _flecs_sparse_add(sparse, elem_size);
 }
 
-int32_t ecs_sparse_last_id(
+uint64_t ecs_sparse_last_id(
     const ecs_sparse_t *sparse)
 {
     return flecs_sparse_last_id(sparse);
