+++ src/sparse.c
@@ -760,7 +760,7 @@ void* _ecs_sparse_add(
     return _flecs_sparse_add(sparse, elem_size);
 }
 
-int32_t ecs_sparse_last_id(
+uint64_t ecs_sparse_last_id(
     const ecs_sparse_t *sparse)
 {
     return flecs_sparse_last_id(sparse);
